Attempt at MATLAB GUI with AppDesigner
Using the built-in App Designer utility, a draft GUI was made (technically prior to the python attempt). Once a majority of the GUI appeared to be working, on testing it the interface appeared to be sluggish, and the selection buttons would somehow 'slip' with respect to hte list of images, causing some very undesireable behaviors. Potentially something wrong with how this GUI handles user input, not interesting enough to continue.